pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["bot*", "agent*", "db*"]

[project]
name = "idea-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "pydantic-settings>=2.0",
    "langgraph>=0.2",
    "langchain-anthropic>=0.2",
    "langchain-openai>=0.2",
    "langchain-google-genai>=2.0",
    "praw>=7.7",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
